requirements.txt
numpy
Pillow
tensorflow-cpu
flask
Werkzeug
gunicorn
opencv-contrib-python-headless







